Correction of controller classes
