--- Services/src/Main/Java/Controller/AdminController.java
@@ -1,139 +1,0 @@
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import java.util.List;
-
-@Service
-@Transactional
-@RestController
-@RequestMapping(path = "/admin")
-public class AdminController {
-
-    @Autowired
-    @PersistenceContext
-    private EntityManager entityManager;
-    
-    @PutMapping(path = "/api/admin/modify/{type}/{id}",consumes = "application/json",produces = "application/json")
-    public ResponseEntity<AdminActions> updateAdminAction(@PathVariable String type, @PathVariable Long id, @RequestBody AdminActions adminAction) {
-        AdminActions existingAction = adminActionsRepository.findById(id).orElse(null);
-        if (existingAction != null) {
-            existingAction.setType(type);
-            AdminActions updatedAction = adminActionsRepository.save(existingAction);
-            return ResponseEntity.ok(updatedAction);
-        } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
-        }
-}
-    @DeleteMapping(path = "/api/admin/delete/{type}/{id}")
-    public ResponseEntity<String> deleteAdminAction(@PathVariable String type, @PathVariable Long id) {
-        AdminActions adminAction = adminActionsRepository.findById(id).orElse(null);
-        if (adminAction != null) {
-            adminActionsRepository.delete(adminAction);
-            return ResponseEntity.ok("Admin action deleted successfully!");
-        } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Admin action not found.");
-        }
-}
-    @GetMapping(path = "/api/admin/logs",produces = "application/json")
-    public ResponseEntity<List<SystemLog>> getAllSystemLogs() {
-        List<SystemLog> systemLogs = systemLogRepository.findAll();
-        return ResponseEntity.ok(systemLogs);
-}
-    @PostMapping(path = "/api/admin/advertisements".consumes = "application/json",produces = "application/json")
-    public ResponseEntity<Advertisement> createAdvertisement(@RequestBody Advertisement advertisement) {
-        Advertisement savedAdvertisement = advertisementService.saveAdvertisement(advertisement);
-        return ResponseEntity.status(HttpStatus.CREATED).body(savedAdvertisement);
-}
-    @PostMapping(value = "/api/admin/buddy/assign",consumes = "application/json",produces = "application/json")
-    public ResponseEntity<String> assignBuddyToInternationalStudent(@RequestBody BuddyAssignmentRequest request) {
-        boolean assignmentSuccessful = buddyAssignmentService.assignBuddy(request.getInternationalStudentId(), request.getBuddyId());
-        if (assignmentSuccessful) {
-            return ResponseEntity.status(HttpStatus.CREATED).body("Buddy assigned successfully!");
-        } else {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Failed to assign buddy.");
-}
-    @PostMapping(path = "/api/admin/users/manage",consumes = "application/json",produces = "application/json")
-    public ResponseEntity<?> createUser(@RequestBody User user) {
-    try {
-        entityManager.persist(user);
-        return ResponseEntity.ok("User created successfully.");
-    } catch (Exception e) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-}
-    @PutMapping(path = "/api/admin/users/manage",consumes = "application/json",produces = "application/json")
-    public ResponseEntity<?> updateUser(@RequestBody User updatedUser) {
-    try {
-        User existingUser = userRepository.findById(updatedUser.getId())
-        .orElseThrow(() -> new ResourceNotFoundException("User not found with ID: " + updatedUser.getId()));
-        existingUser.setSomeProperty(updatedUser.getSomeProperty());
-        userRepository.save(existingUser);
-        return ResponseEntity.ok("User updated successfully.");
-    } catch (ResourceNotFoundException e) {
-        return ResponseEntity.notFound().build();
-    } catch (Exception e) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-}
-    @DeleteMapping(path = "/api/admin/users/manage")
-    public ResponseEntity<?> deleteUser(@RequestBody int userID) {
-    try {
-        Optional<User> optionalUser = userRepository.findById(userID);
-        if (optionalUser.isPresent()) {
-            User user = optionalUser.get();
-            userRepository.delete(user);
-            return ResponseEntity.ok("User deleted successfully.");
-        } else {
-            return ResponseEntity.notFound().build();
-        }
-    } catch (Exception e) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-}
-    @PostMapping(path = "/api/admin/jobs",consumes = "application/json",produces = "application/json")
-    public ResponseEntity<?> createJobAdvertisement(@RequestBody JobAdvertisement jobAdvertisement) {
-    try {
-        entityManager.persist(jobAdvertisement);
-        return ResponseEntity.ok("Job advertisement created successfully.");
-    } catch (Exception e) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-}
-    @DeleteMapping("/api/forum/delete/{postId}")
-    public ResponseEntity<String> deleteForumPost(@PathVariable Long postId) {
-        try {
-            postRepository.deleteByPostId(postId);
-            return ResponseEntity.ok("Forum post deleted successfully");
-        } catch (EmptyResultDataAccessException e) {
-            return ResponseEntity.notFound().build();
-        }
-}
-    @PutMapping(path = "/api/forum/edit/{postId}")
-    public ResponseEntity<ForumPost> updateForumPost(@PathVariable Long postId, @RequestBody ForumPost updatedPost) {
-        ForumPost existingPost = postRepository.findById(postId)
-        .orElseThrow(() -> new ResourceNotFoundException("ForumPost not found with ID: " + postId));
-        ForumPost savedPost = postRepository.save(existingPost);
-        return ResponseEntity.ok(savedPost);
-}
-    @PosttMapping(path = "/api/auth/login/{adminId}",consumes = "application/json",produces = "application/json")
-    @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<User> setRole(@PathVariable("adminId") Long adminId,@RequestParam ERole role) {
-        Optional<User> userOptional = userRepository.findById(adminId);
-        Optional<Role> roleOptional = roleRepository.findByName(role);
-
-        if (userOptional.isPresent() && roleOptional.isPresent()) {
-            User user = userOptional.get();
-            Role assignedRole = roleOptional.get();
-
-            user.getRoles().add(assignedRole);
-            userRepository.save(user);
-
-            return ResponseEntity.ok(user);
-        } else {
-            return ResponseEntity.notFound().build();
-        }
-}
-    public enum ERole {
-    ROLE_STUDENT,
-    ROLE_SUPERADMIN,
-    ROLE_ADMIN
-}
