+++ Services/src/Main/Java/Data/Repository/SuperAdminRepository.java
@@ -1,0 +1,12 @@
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AccountManagementRepository extends JpaRepository<AccountManagement, Long> {
+    // You can add custom query methods here if needed
+}
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    // You can add custom query methods here if needed
+}
+
